[2/3] Settings: make overlay permission check configurable
google added security check when changing app permissions
to check for apps running with enabled draw over apps
(SYSTEM_ALERT_WINDOW) permission.

Since OmniSwitch is such an app it will always trigger this
and requires users to remove permissions before changing
app permissions. After that the permission must be added back.

Since this is a PITA add a config that will disable the check
This is reducing security against tapjacking so make it
clear to the user what it means to disable it

TODO: maybe add a even more prominent alert when
enabling the value

Change-Id: I370a69467eaf1be2063872bbede13693fddeb31a
Signed-off-by: mydongistiny <jaysonedson@gmail.com>