Display the correct default for mobile data always on.
Mobile data always on was enabled by changing the default in
ConnectivityService to 1. However, the default value in the
developer options setting was left at 0. This means that by
default (i.e., when the setting is empty), mobile data always on
is enabled but appears in developer options as disabled.

Fix this by changing the default in developer options. This is
simpler than the other options (adding database upgrade code to
settings, making the default be controlled by config.xml and
ensuring that when a config changed broadcast is received the
value is re-rease, hardcoding a default value as an @hide
constant in Settings, etc.)

Bug: 23113288
Test: setting is off after "adb shell settings delete global mobile_data_always_on"
Change-Id: I27e2c0af7a03e510e640dbcfa281449d00385154
Signed-off-by: mydongistiny <jaysonedson@gmail.com>